continue to clean up AI junk
