+++ doc/RELEASE_NOTES.md
@@ -1,0 +1,3 @@
+## 2025-10-13 Update
+* Changing a device fan motor or speed now disables rules on that device for 3 hours
+* Air quality variables can now be used in rules
