+++ tests/helpers/browser_helpers.py
@@ -4,10 +4,11 @@ Browser test helpers for Playwright-based tests.
 import logging
 import time
 import sqlite3
+import json
 from playwright.sync_api import Page, expect
 import playwright.sync_api
 from app import db, rules_engine
-
+from app import ae200
 
 logger = logging.getLogger(__name__)
